getLocationActivity->getLocation 함수화
