+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".data.getLocationActivity" />
     </application>
 
 </manifest>
